+++ src/App.js
@@ -4,7 +4,7 @@ import {BrowserRouter as Router, Route, Routes} from 'react-router-dom';
 import Navbar from "./entry/main";
 import Signup from "./entry/signup";
 import Login from "./entry/login";
-import Pdash from "./entry/pdash";
+import Pdash from "./owner/pdash";
 import Cdash from "./entry/cdash";
 import Welcome from "./owner/welcome";
 import Free from "./owner/freeslot";
@@ -12,8 +12,8 @@ import Pprof from "./owner/providerprofile";
 import Park from "./owner/addparking";
 import Book from "./owner/fillslots";
 import ViewParking from "./owner/viewpark";
-import Cprof from "./owner/clientprofile";
-import Fpark from "./owner/fetchparking";
+import Cprof from "./client/clientprofile";
+import Fpark from "./client/fetchparking";
 
 function App() {
   return (
@@ -22,7 +22,6 @@ function App() {
         <Route path="/" element={<Navbar/>}></Route>
         <Route path="/signup" element={<Signup/>}></Route>
         <Route path="/login" element={<Login/>}></Route>
-        <Route path="/cdash" element={<Cdash/>}></Route>
         <Route path="/pdash" element={<Pdash/>}>
           <Route path="/pdash/" element={<Welcome/>}></Route>
           <Route path="/pdash/pprof" element={<Pprof/>}></Route>
@@ -31,6 +30,7 @@ function App() {
           <Route path="/pdash/freeslot" element={<Free/>}></Route>
           <Route path="/pdash/viewpark" element={<ViewParking/>}></Route>
         </Route>
+        <Route path="/cdash" element={<Cdash/>}></Route>
         <Route path="/cprof" element={<Cprof/>}></Route>
         <Route path="/cfpark" element={<Fpark/>}></Route>
     </Routes>
