+++ src/owner/css/sidebar.css
